# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trams API
  version: 1.0.0
  description: |
    Public HTTP surface of the Trams event-driven microservices system, served
    by the API Gateway — the only publicly reachable component.

    Two things are worth knowing before reading the endpoints:

    **Registration is asynchronous downstream.** `POST /auth/register` returns
    201 immediately and the resulting notification is produced later, over a
    message broker. It returns 201 even if the broker is completely down: the
    event is committed to a transactional outbox alongside the user and
    published on recovery. Poll `GET /notifications` to observe the effect.

    **The notification surface is read-only.** There is deliberately no endpoint
    that creates a notification. The only way one exists is by consuming a
    domain event from JetStream — if a "send" endpoint existed here, the User
    Service could call it over REST and the event-driven design would be
    decoration.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.trams.example/api/v1
    description: Production (illustrative)
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Current user
      description: >
        Preferred over `/users/{id}`: it takes no client-supplied identifier,
        so there is no ownership check to get wrong.
      operationId: getCurrentUser
      responses:
        '200':
          description: The authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    patch:
      tags: [Users]
      summary: Update profile
      description: |
        Emits `user.updated` carrying `changedFields`, so the consumer decides
        what is worth notifying about.

        Submitting the current value emits nothing — notifying about a
        non-event is noise. An empty object is rejected rather than treated as
        a silent no-op.
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                email: { type: string, format: email, maxLength: 254 }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: That email is already in use
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

    delete:
      tags: [Users]
      summary: Delete account
      description: |
        Emits `user.deleted`; refresh tokens cascade.

        The deletion notification is delivered *after* the row is gone — the
        event payload carries the email and name, so the consumer never has to
        query a deleted record.

        Do not send `content-type: application/json` with no body; an empty
        JSON body is malformed and is rejected with 400.
      operationId: deleteCurrentUser
      responses:
        '204': { description: Account deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/change-password:
    post:
      tags: [Users]
      summary: Change password
      description: |
        Emits `user.password_changed` and revokes all other sessions.

        The current password is required even though the caller is already
        authenticated: a stolen access token alone must not let an attacker
        lock the owner out. Sessions are revoked because if the change is
        prompted by a compromise, leaving the attacker's refresh token live
        would defeat the point.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, minLength: 1 }
                newPassword: { type: string, minLength: 10, maxLength: 200 }
      responses:
        '204': { description: Password changed; other sessions revoked }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Current password incorrect
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /users/{id}:
    get:
      tags: [Users]
      summary: Fetch a user by id
      description: >
        A non-admin requesting another account receives **404, not 403**. A 403
        would confirm the account exists and allow id enumeration.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users:
    get:
      tags: [Users]
      summary: List all users (admin)
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256-signed access token, 15-minute lifetime.

        The gateway holds only the public key, so it can verify a token but
        cannot mint one — deliberately, since it is the internet-facing
        component. Each backend service re-verifies the signature itself rather
        than trusting a header from upstream.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  schemas:
    User:
      type: object
      description: >
        A user as exposed over HTTP. `password_hash` is structurally absent —
        the DTO type omits it, so it cannot leak by accident.
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [user, admin] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
    Problem:
      type: object
      description: RFC 9457 problem document. Branch on `code`, not on prose.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - INVALID_CREDENTIALS
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - UPSTREAM_UNAVAILABLE
            - INTERNAL_ERROR
        detail:
          type: string
          description: >
            Human-readable. For 5xx this is always a generic string — internal
            messages are logged, never returned.
        errors:
          type: array
          description: Field-level detail on validation failures.
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
        correlationId:
          type: string
          description: Quote this in a bug report; it maps to exact log lines.
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://trams.local/errors/validation-error
            title: Validation Failed
            status: 400
            code: VALIDATION_ERROR
            detail: Request validation failed
            errors:
              - path: email
                message: Must be a valid email address
            correlationId: 0b7c1f2e-91aa-4d3c-9f10-2ac4e5b7d8e9
    Unauthorized:
      description: Missing, malformed, expired or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: >
        Not found — or found but not yours. The two are deliberately
        indistinguishable to prevent enumeration.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...